+++ doc/WHATSNEW.md
@@ -3,9 +3,11 @@ Version ??? (unreleased)
 
 - Breaking change: renamed `dbconsole database:shell` script to `dbconsole sql:shell`
 
-- Improved: dbconsole command: `shell` learned option `--database`
+- Improved: dbconsole command: `sql:shell` learned option `--database`
 
 - Improved: dbconsole commands learned option `--output-file`
+
+- Improved: dbconsole commands learned one-letter shortcuts for many options
 
 - Fixed: display of docs in the web interface